Cleanup basket and switch to owner and qty
Owner is a more accurate description of the person associated with a basket as they could be an anonymous user of the service.

Changed qty to quantity in a basket to be consistent with ingredients.

Fixed a minor issue with a basket not saving changes to the db

Re-introduced the basket tests and fixed a few issues